+++ src/Phan/Analysis/ConditionVisitorUtil.php
@@ -202,6 +202,8 @@ trait ConditionVisitorUtil
      * and update the context.
      *
      * Note: It's expected that $should_filter_cb returns false on the new UnionType of that variable.
+     *
+     * @param Node $var_node a node of kind ast\AST_VAR, ast\AST_PROP, or ast\AST_DIM
      */
     final protected function updateVariableWithConditionalFilter(
         Node $var_node,
@@ -216,6 +218,8 @@ trait ConditionVisitorUtil
             if (\is_null($variable)) {
                 if ($var_node->kind === ast\AST_DIM) {
                     return $this->updateDimExpressionWithConditionalFilter($var_node, $context, $should_filter_cb, $filter_union_type_cb, $suppress_issues);
+                } elseif ($var_node->kind === ast\AST_PROP) {
+                    return $this->updatePropertyExpressionWithConditionalFilter($var_node, $context, $should_filter_cb, $filter_union_type_cb, $suppress_issues);
                 }
                 return $context;
             }
@@ -294,12 +298,59 @@ trait ConditionVisitorUtil
         return $context;
     }
 
+    /**
+     * @param Node|mixed $node
+     */
+    protected static function isThisVarNode($node) : bool
+    {
+        return $node instanceof Node && $node->kind === ast\AST_VAR &&
+            $node->children['name'] === 'this';
+    }
+
+    /**
+     * Analyze an expression such as `assert(!is_int($this->prop_name))`
+     * and infer the effects on $this->prop_name in the local scope.
+     *
+     * @param Node $node a node of kind ast\AST_PROP
+     */
+    final protected function updatePropertyExpressionWithConditionalFilter(
+        Node $node,
+        Context $context,
+        Closure $should_filter_cb,
+        Closure $filter_union_type_cb,
+        bool $unused_suppress_issues
+    ) : Context {
+        if (!self::isThisVarNode($node->children['expr'])) {
+            return $context;
+        }
+        $property_name = $node->children['prop'];
+        if (!is_string($property_name)) {
+            return $context;
+        }
+        return $this->modifyPropertyOfThisSimple(
+            $node,
+            static function (UnionType $type) use ($should_filter_cb, $filter_union_type_cb) : UnionType {
+                if (!$should_filter_cb($type)) {
+                    return $type;
+                }
+                return $filter_union_type_cb($type);
+            },
+            $context
+        );
+    }
+
     final protected function updateVariableWithNewType(
         Node $var_node,
         Context $context,
         UnionType $new_union_type,
         bool $suppress_issues
     ) : Context {
+        if ($var_node->kind === ast\AST_PROP) {
+            return $this->modifyPropertySimple($var_node, function (UnionType $unused) use ($new_union_type) : UnionType{
+                return $new_union_type;
+            }, $context);
+        }
+        // TODO: Support ast\AST_DIM
         try {
             // Get the variable we're operating on
             $variable = $this->getVariableFromScope($var_node, $context);
@@ -341,36 +392,15 @@ trait ConditionVisitorUtil
         Context $context = null
     ) : Context {
         $context = $context ?? $this->context;
-        $var_name = $var_node->children['name'] ?? null;
-        // Don't analyze variables such as $$a
-        if (\is_string($var_name) && $var_name) {
-            try {
-                $expr_type = UnionTypeVisitor::unionTypeFromLiteralOrConstant($this->code_base, $context, $expr);
-                if (!$expr_type) {
-                    return $context;
-                }
-                // Get the variable we're operating on
-                $variable = $this->getVariableFromScope($var_node, $context);
-                if (\is_null($variable)) {
-                    return $context;
-                }
-
-                // Make a copy of the variable
-                $variable = clone($variable);
-
-                $variable->setUnionType($expr_type);
-
-                // Overwrite the variable with its new type in this
-                // scope without overwriting other scopes
-                $context = $context->withScopeVariable(
-                    $variable
-                );
+        try {
+            $expr_type = UnionTypeVisitor::unionTypeFromLiteralOrConstant($this->code_base, $context, $expr);
+            if (!$expr_type) {
                 return $context;
-            } catch (\Exception $_) {
-                // Swallow it (E.g. IssueException for undefined variable)
             }
+        } catch (\Exception $_) {
+            return $context;
         }
-        return $context;
+        return $this->updateVariableWithNewType($var_node, $context, $expr_type, true);
     }
 
     /**
@@ -435,7 +465,7 @@ trait ConditionVisitorUtil
     }
 
     /**
-     * @param Node $var_node
+     * @param Node $var_node a node of type ast\AST_VAR, ast\AST_DIM (planned), or ast\AST_PROP
      * @param Node|int|float|string $expr
      * @return Context - Constant after inferring type from an expression such as `if ($x !== 'literal')`
      * @suppress PhanUnreferencedPublicMethod referenced in ConditionVisitorInterface
@@ -446,31 +476,28 @@ trait ConditionVisitorUtil
         Context $context = null
     ) : Context {
         $context = $context ?? $this->context;
-        $var_name = $var_node->children['name'] ?? null;
-        if (\is_string($var_name)) {
-            try {
-                if ($expr instanceof Node) {
-                    if ($expr->kind === ast\AST_CONST) {
-                        $expr_name_node = $expr->children['name'];
-                        if ($expr_name_node->kind === ast\AST_NAME) {
-                            // Currently, only add this inference when we're absolutely sure this is a check rejecting null/false/true
-                            $expr_name = $expr_name_node->children['name'];
-                            switch (\strtolower($expr_name)) {
-                                case 'null':
-                                    return $this->removeNullFromVariable($var_node, $context, false);
-                                case 'false':
-                                    return $this->removeFalseFromVariable($var_node, $context);
-                                case 'true':
-                                    return $this->removeTrueFromVariable($var_node, $context);
-                            }
+        try {
+            if ($expr instanceof Node) {
+                if ($expr->kind === ast\AST_CONST) {
+                    $expr_name_node = $expr->children['name'];
+                    if ($expr_name_node->kind === ast\AST_NAME) {
+                        // Currently, only add this inference when we're absolutely sure this is a check rejecting null/false/true
+                        $expr_name = $expr_name_node->children['name'];
+                        switch (\strtolower($expr_name)) {
+                            case 'null':
+                                return $this->removeNullFromVariable($var_node, $context, false);
+                            case 'false':
+                                return $this->removeFalseFromVariable($var_node, $context);
+                            case 'true':
+                                return $this->removeTrueFromVariable($var_node, $context);
                         }
                     }
-                } else {
-                    return $this->removeLiteralScalarFromVariable($var_node, $context, $expr, true);
                 }
-            } catch (\Exception $_) {
-                // Swallow it (E.g. IssueException for undefined variable)
+            } else {
+                return $this->removeLiteralScalarFromVariable($var_node, $context, $expr, true);
             }
+        } catch (\Exception $_) {
+            // Swallow it (E.g. IssueException for undefined variable)
         }
         return $context;
     }
@@ -594,8 +621,14 @@ trait ConditionVisitorUtil
     {
         '@phan-var ConditionVisitorUtil|ConditionVisitorInterface $this';
         $kind = $var_node->kind;
-        if ($kind === ast\AST_VAR) {
+        if ($kind === ast\AST_VAR || $kind === ast\AST_DIM) {
             return $condition->analyzeVar($this, $var_node, $expr_node);
+        }
+        if ($kind === ast\AST_PROP) {
+            if (self::isThisVarNode($var_node->children['expr']) && is_string($var_node->children['prop'])) {
+                return $condition->analyzeVar($this, $var_node, $expr_node);
+            }
+            return null;
         }
         if ($kind === ast\AST_CALL) {
             $name = $var_node->children['expr']->children['name'] ?? null;
@@ -801,8 +834,8 @@ trait ConditionVisitorUtil
     {
         if (\count($args) >= 1) {
             $arg = $args[0];
-            // Phan also supports `if (!is_array($x['field']))`
-            return ($arg instanceof Node) && (\in_array($arg->kind, [ast\AST_VAR, ast\AST_DIM], true));
+            // Phan also supports `if (!is_array($x['field']))` and `if (!is_array($this->propName))`
+            return ($arg instanceof Node) && (\in_array($arg->kind, [ast\AST_VAR, ast\AST_DIM, ast\AST_PROP], true));
         }
         return false;
     }
@@ -888,6 +921,10 @@ trait ConditionVisitorUtil
         }
         if ($node->kind === ast\AST_DIM) {
             return $this->modifyComplexDimExpression($node, $type_modification_callback, $context, $args);
+        } elseif ($node->kind === ast\AST_PROP) {
+            if (self::isThisVarNode($node->children['expr'])) {
+                return $this->modifyPropertyOfThis($node, $type_modification_callback, $context, $args);
+            }
         }
         return $context;
     }
@@ -897,12 +934,11 @@ trait ConditionVisitorUtil
      * @param Closure(CodeBase,Context,Variable,array<int,mixed>):void $type_modification_callback
      *        A closure acting on a Variable instance (not really a variable) to modify its type
      *
-     *        This is a function such as is_array, is_null (questionable),
+     *        This is a function such as is_array, is_null (questionable), etc.
      * @param Context $context
      * @param array<int,mixed> $args
-     * @return Context
      */
-    protected function modifyComplexDimExpression(Node $node, Closure $type_modification_callback, Context $context, array $args)
+    protected function modifyComplexDimExpression(Node $node, Closure $type_modification_callback, Context $context, array $args) : Context
     {
         $var_name = $this->getVarNameOfDimNode($node->children['expr']);
         if (!is_string($var_name)) {
@@ -928,6 +964,74 @@ trait ConditionVisitorUtil
             $node,
             $new_field_type
         ))->__invoke($node);
+    }
+
+    /**
+     * Return a context with overrides for the type of a property in the local scope.
+     *
+     * @param Node $node a node of kind ast\AST_PROP (e.g. the argument of is_array($this->prop_name))
+     * @param Closure(CodeBase,Context,Variable,array<int,mixed>):void $type_modification_callback
+     *        A closure acting on a Variable instance (not really a variable) to modify its type
+     *
+     *        This is a function such as is_array, is_null, etc.
+     * @param Context $context
+     * @param array<int,mixed> $args
+     */
+    protected function modifyPropertyOfThis(Node $node, Closure $type_modification_callback, Context $context, array $args) : Context
+    {
+        $property_name = $node->children['prop'];
+        if (!is_string($property_name)) {
+            return $context;
+        }
+        // Give the property a type and compute the new type
+        $old_property_type = UnionTypeVisitor::unionTypeFromNode($this->code_base, $context, $node);
+        $property_variable = new Variable($context, "__phan", $old_property_type, 0);
+        $type_modification_callback($this->code_base, $context, $property_variable, $args);
+        $new_property_type = $property_variable->getUnionType();
+        if ($new_property_type->isEqualTo($old_property_type)) {
+            // This didn't change anything
+            return $context;
+        }
+        return $context->withThisPropertySetToTypeByName($property_name, $new_property_type);
+    }
+
+    /**
+     * Return a context with overrides for the type of a property in the local scope.
+     *
+     * @param Node $node a node of kind ast\AST_PROP (e.g. the argument of is_array($this->prop_name))
+     * @param Closure(UnionType):UnionType $type_mapping_callback
+     *        Given a union type, returns the resulting union type.
+     * @param Context $context
+     */
+    protected function modifyPropertyOfThisSimple(Node $node, Closure $type_mapping_callback, Context $context) : Context
+    {
+        $property_name = $node->children['prop'];
+        if (!is_string($property_name)) {
+            return $context;
+        }
+        // Give the property a type and compute the new type
+        $old_property_type = UnionTypeVisitor::unionTypeFromNode($this->code_base, $context, $node);
+        $new_property_type = $type_mapping_callback($old_property_type);
+        if ($new_property_type->isEqualTo($old_property_type)) {
+            // This didn't change anything
+            return $context;
+        }
+        return $context->withThisPropertySetToTypeByName($property_name, $new_property_type);
+    }
+
+    /**
+     * @param Node $node a node of kind ast\AST_PROP (e.g. the argument of is_array($this->prop_name))
+     *                   This is a no-op of the expression is not $this.
+     * @param Closure(UnionType):UnionType $type_mapping_callback
+     *        Given a union type, returns the resulting union type.
+     * @param Context $context
+     */
+    protected function modifyPropertySimple(Node $node, Closure $type_mapping_callback, Context $context) : Context
+    {
+        if (!self::isThisVarNode($node->children['expr'])) {
+            return $context;
+        }
+        return self::modifyPropertyOfThisSimple($node, $type_mapping_callback, $context);
     }
 
     /**