+++ src/Phan/Language/Element/Parameter.php
@@ -153,8 +153,7 @@ class Parameter extends Variable
      */
     public function handleDefaultValueOfNull()
     {
-        $default_value_type = $this->default_value_type;
-        if ($default_value_type && $default_value_type->isType(NullType::instance(false))) {
+        if ($this->default_value_type && $this->default_value_type->isType(NullType::instance(false))) {
             // If it isn't already nullable, convert the parameter type to nullable.
             $this->convertToNullable();
         }